Fix the gosec rule
The version of go tools, including gosec, was not defined in the
Makefile; so the latest version was downloaded. The latest
version of gosec contains a new alert that is breaking our build.
I fixed it.

To avoid the same issue in the future, we are defining a specific
version for all go tools.

It also includes a dependency in the go.mod (import added in the last
commit).

Fixes #17